requirements.txt
asgiref==3.3.1
blinker==1.3
cffi==1.14.4
chardet==2.3.0
cloud-init==19.3
command-not-found==0.3
configobj==5.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.2.1
defusedxml==0.6.0
Django==3.1.4
django-allauth==0.44.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
drf-yasg==1.20.0
idna==2.0
importlib-metadata==3.1.1
inflection==0.5.1
itypes==1.2.0
Jinja2==2.8
jsonpatch==1.10
jsonpointer==1.9
language-selector==0.1
Markdown==3.3.3
MarkupSafe==0.23
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.7
prettytable==0.7.2
psycopg2-binary==2.8.6
pyasn1==0.1.9
pycparser==2.20
pycurl==7.43.0
pygobject==3.20.0
PyJWT==1.7.1
pyparsing==2.4.7
pyserial==3.0.1
python-apt==1.1.0b1+ubuntu0.16.4.5
python-debian==0.1.27
python-systemd==231
python3-openid==3.2.0
pytz==2020.4
PyYAML==3.11
requests==2.9.1
requests-oauthlib==1.3.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
simplejson==3.17.2
six==1.10.0
sqlparse==0.4.1
ssh-import-id==5.5
ufw==0.35
unattended-upgrades==0.1
uritemplate==3.0.1
urllib3==1.13.1
zipp==3.4.0
